==> scripts/01b. Check state population and rental data.ipynb <==
## This notebook contains:
### 01. Import libraries and data
### 02. A note on umlauts
### 03. Check data

#### Note: The source of this data is a spreadsheet of my own design where I combined German state-level population and rental property data. The data came from the Federal Statistical Office of Germany. The population trend variables refer the average rate of population change in each state between 2018 and 2020, when the rental data was collected.


# 01. Import libraries and data

# import libraries
import pandas as pd
import numpy as np
import os

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\german rent'

df_pop = pd.read_csv(os.path.join(path, '02 data', 'original data', 'state_pop.csv'), index_col = False)

# 02. A note on umlauts

Save any Excel files as CSV UTF-8, then import normally and the special characters should be preserved. When exporting, specify encoding = 'utf8' after path and before index (see data import and cleaning)

# 03. Check data

df_pop.head()

df_pop.shape

df_pop.dtypes
# make sure none of the number columns have "object" for data type
