==> scripts/02. Population data merge.ipynb <==
# This notebook contains:
## 01. Import libraries and data
## 02. Merge rent data with state population data
## 03. Merge rent/state pop data with district pop data

# 01. Import libraries and data

# import libraries
import pandas as pd
import numpy as np
import os

# create shortcut for data imports
path = r'../cleaned data/'

# state population data
df_pop = pd.read_csv(os.path.join(path, 'state_pop.csv'), index_col = False)

df_pop.head()

df_pop.shape

# district population data
df_district_pop = pd.read_csv(os.path.join(path, 'district_pop.csv'), index_col = False)

df_district_pop.head()

# rent data
df_rent = pd.read_csv(os.path.join(path, 'rent_cleaned.csv'), index_col = False)

df_rent.shape

df_rent.head(10)

# 02. Merge rent data with state population data

# rename state column so it matches with merge column in the rent data
df_pop = df_pop.rename(columns={'state': 'regio1'})

df_pop.head()

# merge
df_merge = df_rent.merge(df_pop, on = 'regio1', indicator = True)

df_merge.head()

# check merge flag
df_merge['_merge'].value_counts()

# rearrange columns
df_merge = df_merge[['scoutId', 'regio1', 'baseRent', 'totalRent', 'price/unit', 'pricetrend', 'livingSpace', 'geo_plz', 'picturecount', 'newlyConst', 'yearConstructed', 'firingTypes', 'heatingType', 'hasKitchen', 'cellar', 'condition', 'street', 'lift', 'typeOfFlat', 'noRooms', 'floor', 'balcony', 'garden',
'baseRentRange', 'noRoomsRange', 'livingSpaceRange', 'yearConstructedRange', 'regio2', 'regio3', 'description', 'date', 'population', 'population trend', 'pop trend cat', 'surface area', 'population density', 'space/person', 'people/apt', '_merge']]


# drop merge flag
df_merge = df_merge.drop(['_merge'], axis=1)


df_merge.head()

# rename columns to match with naming convention in rent data
df_merge = df_merge.rename(columns={'geo_plz': 'plz', 'population trend': 'populationTrend', 'pop trend cat': 'popTrendCat', 'surface area': 'surfaceArea', 'population density': 'populationDensity'})

df_merge.head()

df_merge.to_csv(r'C:\Users\jacym\Desktop\Career Foundry projects\german rent\02 Data\cleaned data\rent_pop_merged.csv', encoding = 'utf8', index=False)

# 03. Merge rent/state pop data with district pop data

df_merge2 = df_merge.merge(df_district_pop, on = 'regio2', indicator = True)

df_merge2.head()

# drop merge flag column
df_merge2 = df_merge2.drop(['_merge'], axis=1)

df_merge2.columns

# export merged data
df_merge2.to_csv(r'C:\Users\jacym\Desktop\Career Foundry projects\german rent\02 Data\cleaned data\rent_district_pop_merged.csv', encoding = 'utf8', index=False)
