==> scripts/05. Multiple linear regression analysis.ipynb <==
## Multiple linear regression analysis

### This script contains the following:

#### 01. Import libraries and data
#### 02. Clean data
#### 03. Data prep for regression analysis
#### 04. Regression analysis
#### 05. Model performance stats
#### 06. Other model results

### Note: The top-performing model, seen below, incorporated kitchen, district population, district population trend, state population trend, picture count, popdensity, balcony, lift and living space to predict total rent with an R-squared of 72 and mean absolute error of 182. 
### The persistence of the large error and relatively low r-squared score suggests that multiple linear regression isn't the best approach for this dataset.

### 1. Import libraries and data

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib
import matplotlib.pyplot as plt
import os
import sklearn
from sklearn.model_selection import train_test_split 
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error, r2_score

# This option ensures that the graphs you create are displayed within the notebook without the need to "call" them specifically.

%matplotlib inline

path = r'C:\Users\jacym\Desktop\Career Foundry projects\german rent\02 data'

df = pd.read_csv(os.path.join(path, 'cleaned data', 'rent_district_pop_merged.csv'))

### 2. Data cleaning

df.columns

df.head(10)

df.shape

df.columns

df['hasKitchen'] = df['hasKitchen'].astype(int)

df['balcony'] = df['balcony'].astype(int)

df['lift'] = df['lift'].astype(int)

# Create subset with just the factors that interest me for regression
reg_sub = df[['livingSpace', 'totalRent', 'pricetrend', 'populationTrend', 'districtPopTrend', 'districtPop', 'hasKitchen', 'picturecount', 'populationDensity', 'balcony', 'lift']]

##### Missing values

# Check for missing values

reg_sub.isnull().sum()


# drop blank values 
reg_sub_trim = reg_sub.dropna(inplace=False)

reg_sub_trim.isnull().sum()

##### Duplicates check

dups = reg_sub_trim.duplicated()

dups.shape # No dups

### 3. Data prep for regression analysis 

from sklearn import linear_model

X = reg_sub_trim[['populationTrend', 'livingSpace', 'pricetrend', 'districtPopTrend', 'districtPop', 'hasKitchen', 'picturecount', 'populationDensity', 'balcony', 'lift']]
y = reg_sub_trim['totalRent']

y

#Splitting the dataset
from sklearn.model_selection import train_test_split
x_train, x_test, y_train, y_test = train_test_split(X, y, test_size = 0.3, random_state = 100)

mlr = LinearRegression()  
mlr.fit(x_train, y_train)

### 4. Regression analysis

print('Intercept:', mlr.intercept_)
print('Coefficients:')
list(zip(X, mlr.coef_))

# Prediction of test set
y_pred_mlr= mlr.predict(x_test)
# Predicted values
print('Prediction for test set: {}'.format(y_pred_mlr))

mlr_diff = pd.DataFrame({'Actual value': y_test, 'Predicted value': y_pred_mlr})
mlr_diff.head()

## 05. Model performance stats

from sklearn import metrics
meanAbErr = metrics.mean_absolute_error(y_test, y_pred_mlr)
rootMeanSqErr = metrics.mean_squared_error(y_test, y_pred_mlr, squared=False)
print('R squared: {:.2f}'.format(mlr.score(X,y)*100))
print('Mean Absolute Error:', meanAbErr)
print('Root Mean Square Error:', rootMeanSqErr)
# kitchen, district pop, picture count, popdensity, balcony, lift, living space, district pop, pricetrend

# 06. Other model results 

## Living space, district population, picture count, population density, kitchen, balcony, lift

X2 = reg_sub_trim[['livingSpace', 'districtPop', 'hasKitchen', 'picturecount', 'populationDensity', 'balcony', 'lift']]
y2 = reg_sub_trim['totalRent']

#Splitting the dataset
from sklearn.model_selection import train_test_split
x2_train, x2_test, y2_train, y2_test = train_test_split(X2, y2, test_size = 0.3, random_state = 100)

mlr = LinearRegression()  
mlr.fit(x2_train, y2_train)

### Regression analysis

print('Intercept:', mlr.intercept_)
print('Coefficients:')
list(zip(X2, mlr.coef_))

# Prediction of test set
y_pred_mlr= mlr.predict(x2_test)
# Predicted values
print('Prediction for test set: {}'.format(y_pred_mlr))

mlr_diff = pd.DataFrame({'Actual value': y2_test, 'Predicted value': y_pred_mlr})
mlr_diff.head()

from sklearn import metrics
meanAbErr = metrics.mean_absolute_error(y2_test, y_pred_mlr)
rootMeanSqErr = metrics.mean_squared_error(y2_test, y_pred_mlr, squared=False)
print('R squared: {:.2f}'.format(mlr.score(X2,y2)*100))
print('Mean Absolute Error:', meanAbErr)
print('Root Mean Square Error:', rootMeanSqErr)
# kitchen, district pop, picture count, popdensity, balcony, lift, living space

## Living space, district population trend and district population

X3 = reg_sub_trim[['livingSpace', 'districtPop', 'districtPopTrend']]
y3 = reg_sub_trim['totalRent']

#Splitting the dataset
from sklearn.model_selection import train_test_split
x3_train, x3_test, y3_train, y3_test = train_test_split(X3, y3, test_size = 0.3, random_state = 100)

mlr = LinearRegression()  
mlr.fit(x3_train, y3_train)

### Regression analysis

print('Intercept:', mlr.intercept_)
print('Coefficients:')
list(zip(X3, mlr.coef_))

# Prediction of test set
y_pred_mlr= mlr.predict(x3_test)
# Predicted values
print('Prediction for test set: {}'.format(y_pred_mlr))

mlr_diff = pd.DataFrame({'Actual value': y3_test, 'Predicted value': y_pred_mlr})
mlr_diff.head()

from sklearn import metrics
meanAbErr = metrics.mean_absolute_error(y3_test, y_pred_mlr)
rootMeanSqErr = metrics.mean_squared_error(y3_test, y_pred_mlr, squared=False)
print('R squared: {:.2f}'.format(mlr.score(X3,y3)*100))
print('Mean Absolute Error:', meanAbErr)
print('Root Mean Square Error:', rootMeanSqErr)
# livingSpace, district population trend and district population
