==> scripts/07. Analyzing the clusters.ipynb <==
## This notebook contains:
### 01. Import libraries and data
### 02. Calculate descriptive statistics for each cluster
### 03. View cluster counts by location
### 04. Assess other cluster qualities- kitchen, flat type, state population category
### 05. See how many good deals there are in each cluster

# 01. Import libraries and data

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib
import os

# This command propts matplotlib visuals to appear in the notebook 

%matplotlib inline

path = r'C:\Users\jacym\Desktop\Career Foundry projects\german rent'

# Import rent data

df = pd.read_csv(os.path.join(path, '02 data', 'cleaned data', 'clustered_data.csv'), index_col = False)

df.head()

df.shape

# count units in each cluster
df.groupby(['cluster']).agg({'scoutId': ['nunique']})

# 02. Calculate descriptive statistics for each cluster

df.groupby('cluster').agg({'totalRent':['mean', 'std'], 
                         'price/unit':['mean', 'std'],
                           'livingSpace':['mean', 'std']
                          })

stats = df.groupby('cluster').agg({'totalRent':['mean', 'std'], 
                         'price/unit':['mean', 'std'],
                           'livingSpace':['mean', 'std']
                          })
stats.to_clipboard()

# 03. View cluster counts by location

df.groupby(['cluster', 'regio1']).agg({'scoutId': ['nunique']})

# copy to clipboard to go over locations in Excel
cluster_locs = df.groupby(['cluster', 'regio1']).agg({'scoutId': ['nunique']})
cluster_locs.to_clipboard()

# 04. Assess other cluster qualities: kitchen, flat type, state population category

df.groupby(['cluster', 'hasKitchen']).agg({'hasKitchen': ['count']})

Most likely to have a kitchen: Upscale units. Least likely: Budget.

df.groupby(['cluster', 'typeOfFlat'], dropna = True).agg({'typeOfFlat': ['count']})

df.groupby(['cluster', 'popTrendCat']).agg({'popTrendCat': ['count']})

Big city units are all in states with rising populations. 
Midlow units are mostly in states with stable populations.
Upscale units are mostly in states with increasing or stable populations.
Budget units are mostly in states with decreasing populations.
Midhigh units are the most spread out.

# 05. See how many good deals there are in each cluster

upscale = df[df['cluster'] == 'upscale']
bigcity = df[df['cluster'] == 'bigcity']
midhigh = df[df['cluster'] == 'midhigh']
midlow = df[df['cluster'] == 'midlow']
budget = df[df['cluster'] == 'budget']


df['cluster'].value_counts()

I'm defining a good deal as ana apartment that is at least half a standard deviation below the average total rent for that cluster. To see how many good deals there are in each cluster, search for rows in each subset with totalRent less than or equal to the good deal threshold. 

df.groupby('cluster').agg({'totalRent':['mean', 'std']})

upscale[upscale['totalRent'] <= 1221.01]

bigcity[bigcity['totalRent'] <= 947.15]

midhigh[midhigh['totalRent'] <= 608.49]

midlow[midlow['totalRent'] <= 570.94]

budget[budget['totalRent'] <= 402.34]
